Remove rarity from layer class
Rarity of layer is the probability for a layer to appear. In my example, an NFT doesn't necessary have an accessory.
This feature has been replaced by the ability to paste automatically a blank picture if the addition of all entities indicated by the user is under 100%.
e.g.:
blue (default 50%) : 45
green (default 50%) : 45
The sum of probabilities isn't equal to 100%, should we replace the missing percentage by nothing? (y/n) (Default: : y): y